requirements.txt

Pillow

# Core utilities
coolname
pygments
omegaconf
hydra-core
argdantic
pydantic
pydantic-core
flash-attn
safetensors


# Numerical and scientific stack
numba
einops
adam-atan2-pytorch
bitsandbytes
scikit-learn
tqdm

# AI & model management
wandb
datasets
huggingface-hub
opencv-python
gradio
pylibcudf-cu12
cudf-cu12
lz4